requirements.txt
mlflow
azureml-core
azureml-mlflow
torch
torchaudio
transformers
numpy
pandas
scikit-learn
matplotlib
pytest